refactor: 즐겨찾기 삭제 시, 해당 즐겨찾기 존재 여부 먼저 확인하도록 수정 - 삭제할 리소스에 대한 확인 및 Exception 처리 - 이후에 삭제 가능 여부를 확인 및 Exception 처리하도록 로직 수정

+++ src/main/java/nextstep/subway/application/service/FavoriteService.java
@@ -42,11 +42,13 @@ public class FavoriteService {
                 .collect(Collectors.toList());
     }
 
-    public void deleteFavorite(String email, Long id) {
-        findFavorites(email).stream()
-                .filter(favorite -> id.equals(favorite.getId()))
-                .findAny()
-                .orElseThrow(() -> new EntityNotFoundException("즐겨찾기가 존재하지 않습니다."));
+    public void deleteFavorite(Long id, Long memberId) {
+        Favorite favorite = favoriteRepository.findById(id)
+                        .orElseThrow(() -> new EntityNotFoundException("즐겨찾기가 존재하지 않습니다."));
+
+        if(favorite.notMemberEquals(memberId)) {
+            throw new AuthenticationException("삭제할 권한이 없습니다.");
+        }
 
         favoriteRepository.deleteById(id);
     }
